Moved qualitymetrics
